refactor: split trade_parallelism trades into their own testfile
